requirements.txt
absl-py==2.3.1
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.12
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.7
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.0
filelock==3.19.1
Flask==3.1.0
flatbuffers==25.9.23
fonttools==4.56.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.14.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.11.3
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
nodemon==0.0.17
numpy==2.2.3
opt_einsum==3.4.0
optree==0.17.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pgxnclient==1.3.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==6.32.1
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pure_eval==0.2.3
py4j==0.10.9.7
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.1
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
rich==14.1.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.30.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.5.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.11.0
wrapt==1.17.3
wsproto==1.2.0
